requirements.txt
alpaca-trade-api==3.0.2
python-dotenv==1.0.0
pandas==2.1.4
numpy==1.26.2
ta==0.10.2
scikit-learn==1.3.2 